requirements.txt
#api
fastapi
uvicorn[standard]
gunicorn
python-multipart

#orm
sqlalchemy
alembic
psycopg2-binary

#data
pandas
python-dotenv

#auth
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]

#dash
prometheus-fastapi-instrumentator
streamlit

passlib
bcrypt<4.0